+++ src/sql/connection.py
@@ -360,7 +360,6 @@ class Connection:
 
         if descriptor:
             is_custom_connection_ = Connection.is_custom_connection(descriptor)
-
             if isinstance(descriptor, Connection):
                 cls.current = descriptor
             elif isinstance(descriptor, Engine):
@@ -369,7 +368,6 @@ class Connection:
                 cls.current = CustomConnection(descriptor, alias=alias)
             else:
                 existing = rough_dict_get(cls.connections, descriptor)
-
                 # NOTE: I added one indentation level, otherwise
                 # the "existing" variable would not exist if
                 # passing an engine object as descriptor.
@@ -377,12 +375,20 @@ class Connection:
                 # is that we're missing some unit tests
                 # when descriptor is a connection object
                 # http://docs.sqlalchemy.org/en/rel_0_9/core/engines.html#custom-dbapi-connect-arguments # noqa
-                cls.current = existing or Connection.from_connect_str(
-                    connect_str=descriptor,
-                    connect_args=connect_args,
-                    creator=creator,
-                    alias=alias,
-                )
+                # if same alias found
+                if existing and existing.alias == alias:
+                    cls.current = existing
+                # if just switching connections
+                elif existing and alias is None:
+                    cls.current = existing
+                # if new alias connection
+                elif existing is None or existing.alias != alias:
+                    cls.current = Connection.from_connect_str(
+                        connect_str=descriptor,
+                        connect_args=connect_args,
+                        creator=creator,
+                        alias=alias,
+                    )
 
         else:
             if cls.connections:
@@ -397,7 +403,6 @@ class Connection:
                 )
             else:
                 raise cls._error_no_connection()
-
         return cls.current
 
     @classmethod