+++ src/tests/test_connection.py
@@ -377,3 +377,27 @@ def test_close_all(ip_empty):
         connections_copy["duckdb://"].execute("").fetchall()
 
     assert not Connection.connections
+
+
+@pytest.mark.parametrize(
+    "old_alias, new_alias",
+    [
+        (None, "duck1"),
+        ("duck1", "duck2"),
+        (None, None),
+    ],
+)
+def test_new_connection_with_alias(ip_empty, old_alias, new_alias):
+    """Test if a new connection with the same url but a
+    new alias is registered for different cases of old alias
+    """
+    ip_empty.run_cell(f"%sql duckdb:// --alias {old_alias}")
+    ip_empty.run_cell(f"%sql duckdb:// --alias {new_alias}")
+    table = ip_empty.run_cell("sql --connections").result
+    if old_alias is None and new_alias is None:
+        assert new_alias not in table
+    else:
+        connection = table[new_alias]
+        assert connection
+        assert connection.url == "duckdb://"
+        assert connection == connection.current